Tablas de referencia empresa, centro costo y plan de cuentas - produccion

+++ database/migrations/2016_05_10_170230_create_plancuenta_table.php
@@ -23,7 +23,7 @@ class CreatePlancuentaTable extends Migration
             $table->integer('plancuentas_centro')->unsigned()->nullable();
             $table->boolean('plancuentas_tercero')->default(false);
             $table->string('plancuentas_tipo', 1);
-            $table->float('plancuentas_tasa')->nullable();
+            $table->double('plancuentas_tasa')->nullable();
             
             $table->integer('plancuentas_nivel1');
             $table->integer('plancuentas_nivel2');
